Fixed compilation problem in linux

+++ src/algoapi/QuotationData.h
@@ -37,7 +37,7 @@ namespace AlgoApi {
     typedef double  d_price_t;
     typedef double  d_amount_t;
     typedef int64_t volume_t;
-    typedef int32_t timestamp_t;
+    typedef int64_t timestamp_t;            /**< timestamp in ms */
 #endif // DEF_TRADE_TYPES
 
     typedef quotation::Stock *PStockData;
